percent-configure: Add
Answers the question "what would %configure do". This is more or less
the output of 'rpm --eval %configure', but that's difficult to invoke on
its own. In addition:

- build and host are $(arch) instead
- you can append options
- $con if set points to the configure script to run

As in:

~/git/xserver/build% con=../autogen.sh percent-configure

Probably this only works on Fedorae newer than whichever version I
generated this from, which I have long since forgotten.